+++ Source/DataSources/ModelGraphics.js
@@ -57,6 +57,7 @@ define([
      * @param {Property} [options.colorBlendAmount=0.5] A numeric Property specifying the color strength when the <code>colorBlendMode</code> is <code>MIX</code>. A value of 0.0 results in the model's rendered color while a value of 1.0 results in a solid color, with any value in-between resulting in a mix of the two.
      * @param {Property} [options.clippingPlanes] A property specifying the {@link ClippingPlaneCollection} used to selectively disable rendering the model.
      * @param {Property} [options.imageBasedLightingFactor=new Cartesian2(1.0, 1.0)] A property specifying the contribution from diffuse and specular image-based lighting.
+     * @param {Property} [options.lightColor] A property specifying the light color to use when shading the model. The default sun light color will be used when <code>undefined</code>.
      *
      * @see {@link https://cesiumjs.org/tutorials/3D-Models-Tutorial/|3D Models Tutorial}
      * @demo {@link https://cesiumjs.org/Cesium/Apps/Sandcastle/index.html?src=3D%20Models.html|Cesium Sandcastle 3D Models Demo}
@@ -99,6 +100,8 @@ define([
         this._clippingPlanesSubscription = undefined;
         this._imageBasedLightingFactor = undefined;
         this._imageBasedLightingFactorSubscription = undefined;
+        this._lightColor = undefined;
+        this._lightColorSubscription = undefined;
         this._definitionChanged = new Event();
 
         this.merge(defaultValue(options, defaultValue.EMPTY_OBJECT));
@@ -272,7 +275,14 @@ define([
          * @memberof ModelGraphics.prototype
          * @type {Property}
          */
-        imageBasedLightingFactor : createPropertyDescriptor('imageBasedLightingFactor')
+        imageBasedLightingFactor : createPropertyDescriptor('imageBasedLightingFactor'),
+
+        /**
+         * A property specifying the {@link Color} of the light source when shading the model.
+         * @memberOf ModelGraphics.prototype
+         * @type {Property}
+         */
+        lightColor : createPropertyDescriptor('lightColor')
     });
 
     /**
@@ -304,6 +314,7 @@ define([
         result.colorBlendAmount = this.colorBlendAmount;
         result.clippingPlanes = this.clippingPlanes;
         result.imageBasedLightingFactor = this.imageBasedLightingFactor;
+        result.lightColor = this.lightColor;
 
         return result;
     };
@@ -339,6 +350,7 @@ define([
         this.colorBlendAmount = defaultValue(this.colorBlendAmount, source.colorBlendAmount);
         this.clippingPlanes = defaultValue(this.clippingPlanes, source.clippingPlanes);
         this.imageBasedLightingFactor = defaultValue(this.imageBasedLightingFactor, source.imageBasedLightingFactor);
+        this.lightColor = defaultValue(this.lightColor, source.lightColor);
 
         var sourceNodeTransformations = source.nodeTransformations;
         if (defined(sourceNodeTransformations)) {